+++ lib/resources/asset_paths.dart
@@ -1,10 +1,6 @@
 const basePath = "lib/resources";
 const _graphics = "$basePath/graphics";
-const _data = "$basePath/data";
 
 const largeLogo = "$_graphics/3.0x/logo.png";
 const smallLogo = "$_graphics/logo.png";
-const colors = "$_data/colors.json";
-const colorSuggestions = "$_data/color_suggestions.json";
-const palettes = "$_data/palettes.json";
-const paletteSuggestions = "$_data/palette_suggestions.json";
+const githubLogo = "$_graphics/github_light.png";