+++ lib/sources/data/color_helpers/converter/color_converter.dart
@@ -5,12 +5,17 @@ import 'package:flutter/foundation.dart';
 class ColorConverter implements Converter {
   const ColorConverter({
     @required List<Converter> converters,
-  }) : _converters = converters;
+  })  : assert(converters != null),
+        _converters = converters;
 
   final List<Converter> _converters;
 
   @override
   Map<Format, String> convert(int r, int g, int b) {
+    if (_converters.isEmpty) {
+      return {};
+    }
+
     final formats = _converters.map((c) => c.convert(r, g, b)).toList();
     final formatData = formats.reduce((map, nextMap) => map..addAll(nextMap));
     return formatData;