+++ lib/sources/domain/view_models/converter/converter_state.dart
@@ -1,12 +1,11 @@
 import 'package:collection/collection.dart';
 import 'package:colored/sources/domain/data_models/format.dart';
 import 'package:flutter/material.dart';
-import 'package:vector_math/hash.dart';
 
 class ConverterState {
   const ConverterState({
     @required this.formatData,
-  });
+  }) : assert(formatData != null);
 
   final Map<Format, String> formatData;
 
@@ -16,11 +15,8 @@ class ConverterState {
       const MapEquality().equals(formatData, other.formatData);
 
   @override
-  int get hashCode => hashObjects([this, formatData]);
+  int get hashCode => formatData.hashCode;
 
   @override
-  String toString() => """ConverterState(
-      formatData: $formatData,
-   )""";
+  String toString() => "$runtimeType(formatData: $formatData)";
 }
-