+++ lib/sources/domain/view_models/displayed_formats/displayed_formats_state.dart
@@ -1,9 +1,8 @@
 import 'package:collection/collection.dart';
 import 'package:colored/sources/domain/data_models/format.dart';
-import 'package:vector_math/hash.dart';
 
 class DisplayedFormatsState {
-  const DisplayedFormatsState(this.formats);
+  const DisplayedFormatsState(this.formats) : assert(formats != null);
 
   final List<Format> formats;
 
@@ -13,8 +12,8 @@ class DisplayedFormatsState {
       const ListEquality().equals(formats, other.formats);
 
   @override
-  int get hashCode => hashObjects([this, formats]);
+  int get hashCode => formats.hashCode;
 
   @override
-  String toString() => "DisplayedFormatsState($formats)";
+  String toString() => "$runtimeType($formats)";
 }