+++ test/unit/sources/domain/view_models/suggestions/color_suggestions/color_suggestions_view_model_test.dart
@@ -80,6 +80,12 @@ void main() {
       });
     });
 
+    group("when stateStream is called", () {
+      test("the stream of the provided state controller is retrieved", () {
+        expect(viewModel.stateStream, stateController.stream);
+      });
+    });
+
     group("when init is called", () {
       test("then ColorSuggestionsFound state is retrieved", () async {
         stateController.stream.listen((event) {